fix(projects): whole-card click, uniform card height, visibility in card edit
- Project cards: the whole card is one click target (stretched-link via the
  title's ::after) with the actions menu kept clickable above it; the description
  reserves a fixed two lines so cards keep a uniform height regardless of
  description length or absence.
- Extract the unified Edit-project dialog (name/description/icon + visibility +
  teams) into a shared component that fetches the project detail by id, and use
  it from both the projects list cards and the project page — so 'Edit details'
  from a card can now change the visibility scope (the list's old dialog only
  edited description). Removes the duplicated list-only dialog.
